sbi: move state extension to its own file
Continue to reduce clutter in sbi.rs.
